pyproject.toml
[tool.poetry]
name = "vkrm-code"
version = "0.1.0"
description = ""
authors = ["Alexander Litvinov <aalitvinov_2@edu.hse.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "vkrm_code"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.3"
pandas = "^2.0.1"
matplotlib = "^3.7.1"
tqdm = "^4.65.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
pytest-playwright = "^0.3.3"
regex = "^2023.5.5"
openpyxl = "^3.1.2"
seaborn = "^0.12.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
